# services/marketing-ops/openapi/marketing-ops.v1.yaml
openapi: 3.1.0
info:
  title: Nexus Marketing Ops API
  version: 1.0.0
  description: REST contract for the operational campaign workspace.
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /campaign-items:
    get:
      operationId: listProductionSchedule
      summary: List the canonical production schedule
      parameters:
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - { name: campaignId, in: query, schema: { type: string, format: uuid } }
        - { name: kind, in: query, schema: { $ref: '#/components/schemas/ItemKind' } }
        - { name: channel, in: query, schema: { $ref: '#/components/schemas/CampaignChannel' } }
        - { name: assigneeId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/ItemStatus' } }
        - { name: priority, in: query, schema: { $ref: '#/components/schemas/ItemPriority' } }
        - { name: cursor, in: query, schema: { type: string, minLength: 1, maxLength: 1024 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        '200':
          description: Authorized items from the canonical query
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductionSchedulePage' }
        default: { $ref: '#/components/responses/Error' }
    post:
      operationId: createProductionItem
      summary: Create a production item
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductionItemCreate' }
      responses:
        '201':
          description: Production item created
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductionItemEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /campaign-items/{itemId}:
    get:
      operationId: getProductionItem
      summary: Read a production item
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
      responses:
        '200':
          description: Production item detail
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductionItemEnvelope' }
        default: { $ref: '#/components/responses/Error' }
    patch:
      operationId: updateProductionItem
      summary: Update production item fields
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductionItemPatch' }
      responses:
        '200':
          description: Production item updated
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductionItemEnvelope' }
        default: { $ref: '#/components/responses/Error' }
  /campaign-items/{itemId}/transition:
    post:
      operationId: transitionProductionItem
      summary: Move a production item through the Phase 3 state machine
      parameters:
        - { $ref: '#/components/parameters/ItemId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductionItemTransition' }
      responses:
        '200':
          description: Production item status changed
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductionItemEnvelope' }
        default: { $ref: '#/components/responses/Error' }
...
  /campaigns/{campaignId}/items:
    post:
      operationId: createCampaignItem
      summary: Create a legacy campaign item
      deprecated: true
      description: Compatibility endpoint retained from Phase 1; item expansion is outside Phase 2.
      parameters:
        - { $ref: '#/components/parameters/CampaignId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CampaignItemCreate' }
      responses:
        '201':
          description: Item draft created
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        default: { $ref: '#/components/responses/Error' }
  /campaigns/{campaignId}/items/{itemId}:
    patch:
      operationId: updateCampaignItem
      summary: Update a legacy campaign item
      deprecated: true
      description: Compatibility endpoint retained from Phase 1; item expansion is outside Phase 2.
      parameters:
        - { $ref: '#/components/parameters/CampaignId' }
        - { $ref: '#/components/parameters/ItemId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CampaignItemPatch' }
      responses:
        '200':
          description: Item draft updated
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  headers:
    ETag:
      description: Quoted current resource or aggregate version.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
  parameters:
...
    CampaignId: { name: campaignId, in: path, required: true, schema: { type: string, format: uuid } }
    ItemId: { name: itemId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, pattern: '^(W/)?("[1-9][0-9]*"|[1-9][0-9]*)$' } }
  responses:
    Error:
      description: Stable public error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    PublicErrorCode:
      type: string
      enum:
        - validation_error
        - unauthorized
        - auth_unavailable
        - tenant_required
        - forbidden
        - not_found
        - origin_forbidden
        - feature_disabled
        - idempotency_key_required
        - precondition_required
        - payload_too_large
        - idempotency_conflict
        - idempotency_in_progress
        - invalid_transition
        - campaign_requirements_missing
        - campaign_read_only
        - primary_owner_required
        - participant_role_invalid
        - participant_exists
        - participant_not_found
        - reference_not_verified
        - artifact_not_owned
        - artifact_not_found
        - artifact_compensation_failed
        - material_type_not_allowed
        - material_too_large
        - material_empty
        - material_exists
        - material_not_found
        - version_conflict
        - invalid_item_transition
        - item_requirements_missing
        - item_content_required
        - item_blocked
        - item_terminal
        - assignee_not_authorized
        - dependency_exists
        - dependency_not_found
        - dependency_self
        - dependency_cycle
        - dependency_campaign_mismatch
        - dependency_terminal
        - artifact_exists
        - artifact_asset_mismatch
        - artifact_link_not_found
        - content_hash_mismatch
        - dependency_unavailable
        - dependency_invalid_response
        - rate_limited
        - internal_error
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, correlationId]
          properties:
            code: { $ref: '#/components/schemas/PublicErrorCode' }
            message: { type: string }
            correlationId: { type: string, minLength: 1, maxLength: 128 }
            details: {}
...
    CampaignChannel:
      type: string
      enum: [email, instagram, linkedin, facebook, whatsapp, website, paid_media, events, press, other]
...
    Page:
      type: object
      additionalProperties: false
      required: [limit, count, nextCursor]
      properties:
        limit: { type: integer, minimum: 1, maximum: 100 }
        count: { type: integer, minimum: 0, maximum: 100 }
        nextCursor: { type: [string, 'null'] }
...
    ItemKind:
      type: string
      enum: [task, email, whatsapp, post, creative, review, milestone]
    ItemStatus:
      type: string
      enum: [draft, ready, in_review, completed, cancelled]
    ItemPriority:
      type: string
      enum: [low, normal, high, urgent]
    ProductionItemEditableProperties:
      type: object
      properties:
        kind: { $ref: '#/components/schemas/ItemKind' }
        title: { type: string, minLength: 1, maxLength: 200 }
        assigneeUserId: { type: [string, 'null'], format: uuid }
        priority: { $ref: '#/components/schemas/ItemPriority' }
        channel:
          oneOf: [{ $ref: '#/components/schemas/CampaignChannel' }, { type: 'null' }]
        description: { type: [string, 'null'], maxLength: 10000 }
        startsAt: { type: [string, 'null'], format: date-time }
        dueAt: { type: [string, 'null'], format: date-time }
        metadata: { type: object, maxProperties: 1000 }
    ProductionItemCreate:
      allOf:
        - { $ref: '#/components/schemas/ProductionItemEditableProperties' }
        - type: object
          required: [campaignId, kind, title]
          properties:
            campaignId: { type: string, format: uuid }
      unevaluatedProperties: false
    ProductionItemPatch:
      allOf:
        - { $ref: '#/components/schemas/ProductionItemEditableProperties' }
      minProperties: 1
      unevaluatedProperties: false
    ProductionItemTransition:
      type: object
      additionalProperties: false
      required: [to]
      properties:
        to: { $ref: '#/components/schemas/ItemStatus' }
    ProductionItem:
      allOf:
        - { $ref: '#/components/schemas/ProductionItemEditableProperties' }
        - type: object
          required:
            - id
            - tenantId
            - campaignId
            - kind
            - title
            - assigneeUserId
            - priority
            - channel
            - description
            - startsAt
            - dueAt
            - metadata
            - content
            - status
            - version
            - createdBy
            - updatedBy
            - createdAt
            - updatedAt
            - completedAt
            - cancelledAt
          properties:
            id: { type: string, format: uuid }
            tenantId: { type: string, format: uuid }
            campaignId: { type: string, format: uuid }
            content: {}
            status: { $ref: '#/components/schemas/ItemStatus' }
            version: { type: integer, minimum: 1 }
            createdBy: { type: string, format: uuid }
            updatedBy: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
            completedAt: { type: [string, 'null'], format: date-time }
            cancelledAt: { type: [string, 'null'], format: date-time }
    ProductionItemEnvelope:
      type: object
      additionalProperties: false
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/ProductionItem' }
    ProductionScheduleItem:
      allOf:
        - { $ref: '#/components/schemas/ProductionItem' }
        - type: object
          required: [campaignName, effectiveAt, isOverdue, isBlocked]
          properties:
            campaignName: { type: string, minLength: 1, maxLength: 200 }
            effectiveAt: { type: [string, 'null'], format: date-time }
            isOverdue: { type: boolean }
            isBlocked: { type: boolean }
    ProductionSchedulePage:
      type: object
      additionalProperties: false
      required: [data, page, meta]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/ProductionScheduleItem' }
        page: { $ref: '#/components/schemas/Page' }
        meta:
          type: object
          additionalProperties: false
          required: [timeZone]
          properties:
            timeZone: { type: string, minLength: 1, maxLength: 100 }
...
    CampaignItemCreate:
      type: object
      additionalProperties: false
      required: [kind, content]
      properties:
        kind: { type: string, minLength: 1, maxLength: 80 }
        title: { type: string, maxLength: 200 }
        content: {}
    CampaignItemPatch:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        title: { type: string, maxLength: 200 }
        content: {}
...